# openapi.yaml
openapi: 3.0.3

# ============================================================================
# 2주차 API 명세 — 인증 (회원가입 + 로그인)
#
# ⚠️ CORS 정책 (필수)
#   Access-Control-Allow-Origin: http://localhost:3000
#   Access-Control-Allow-Methods: GET, POST, PATCH, PUT, DELETE, OPTIONS
#   Access-Control-Allow-Headers: Content-Type, Authorization
#
# ⚠️ 에러 응답 포맷 (필수)
#   { "error": "사람이 읽을 수 있는 에러 메시지" }
#
# ⚠️ 날짜/시간 포맷 (필수)
#   ISO 8601, UTC(Z) 형식: 2026-08-05T10:00:00Z
# ============================================================================

info:
  title: 백엔드 스터디 API - 2주차
  description: |
    2주차: 회원가입 + 비밀번호 해싱, 로그인 + JWT 발급
  version: "2.0.0"

servers:
  - url: http://localhost:8083
    description: 예시 - Go(Echo) 구현체
  - url: http://localhost:8080
    description: 예시 - Spring Boot 구현체
  - url: http://localhost:{port}
    description: 로컬 - 포트를 직접 입력하세요
    variables:
      port:
        default: "8090"
        description: 본인 백엔드가 실행 중인 포트

tags:
  - name: Auth
    description: |
      학습 포인트: 사용자 등록, 비밀번호 보안 저장(해싱), JWT 토큰 생성/검증,
      인증 토큰 기반 세션 관리를 연습합니다.

paths:
  # --------------------------------------------------------------------------
  # 인증 — 회원가입 + 비밀번호 해싱, 로그인 + JWT 발급
  # --------------------------------------------------------------------------
  /auth/register:
    post:
      tags: [Auth]
      summary: 회원가입
      description: |
        학습 포인트: 사용자 등록, 비밀번호 보안 저장(해싱), 중복 검사, 입력 검증을 연습합니다.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            examples:
              default:
                value:
                  email: "user@example.com"
                  password: "securePassword123"
                  name: "홍길동"
      responses:
        "201":
          description: 생성된 사용자 정보
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: 이미 등록된 이메일
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "이미 등록된 이메일입니다."
        "500":
          $ref: "#/components/responses/InternalError"

  /auth/login:
    post:
      tags: [Auth]
      summary: 로그인
      description: |
        학습 포인트: 인증 토큰 기반 인증, JWT 생성/검증, 토큰 기반 세션 관리를 연습합니다.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            examples:
              default:
                value:
                  email: "user@example.com"
                  password: "securePassword123"
      responses:
        "200":
          description: JWT 토큰 발급
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: 잘못된 이메일 또는 비밀번호
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "이메일 또는 비밀번호가 올바르지 않습니다."
        "500":
          $ref: "#/components/responses/InternalError"

components:
  schemas:
    User:
      type: object
      required: [id, email, name, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          example: "user@example.com"
        name:
          type: string
          example: "홍길동"
        createdAt:
          type: string
          format: date-time
          description: ISO 8601, UTC
        updatedAt:
          type: string
          format: date-time
          description: ISO 8601, UTC

    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          example: "securePassword123"
        name:
          type: string
          minLength: 1
          example: "홍길동"

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          example: "securePassword123"

    UserResponse:
      type: object
      required: [id, email, name, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          description: JWT 액세스 토큰
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          $ref: "#/components/schemas/UserResponse"

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "요청한 리소스를 찾을 수 없습니다."

  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: 서버 내부 오류
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"